added minor changes

+++ src/main/java/ru/lazarenko/warehouse/repository/ProductRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import ru.lazarenko.warehouse.entity.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,6 +14,6 @@ public interface ProductRepository extends JpaRepository<Product, Integer> {
     List<Product> findAllByCategoryId(Integer id);
 
     @Query(value = "select p from Product p where p.price >= :min and p.price <= :max")
-    List<Product> findProductsByMinAndMaxPrice(Integer min, Integer max);
+    List<Product> findProductsByMinAndMaxPrice(BigDecimal min, BigDecimal max);
 
 }
